requirements.txt
# requirements.txt

requests
gTTS
playsound
streamlit
pymysql
pandas
python-dotenv
